UHF-12050: Reverted field logic
