+++ modules/helfi_tpr_config/src/Entity/Service.php
@@ -7,29 +7,24 @@ namespace Drupal\helfi_tpr_config\Entity;
 use Drupal\Core\Entity\EntityTypeInterface;
 use Drupal\Core\Field\BaseFieldDefinition;
 use Drupal\Core\StringTranslation\TranslatableMarkup;
-use Drupal\helfi_recommendations\Entity\SuggestedTopicsInterface;
-use Drupal\helfi_recommendations\Entity\SuggestedTopicsTrait;
 use Drupal\helfi_tpr\Entity\Service as BaseService;
 
 /**
  * A bundle class override for Service entities.
  */
-class Service extends BaseService implements SuggestedTopicsInterface {
-
-  use SuggestedTopicsTrait;
+class Service extends BaseService {
 
   /**
    * {@inheritdoc}
    */
   public static function baseFieldDefinitions(EntityTypeInterface $entity_type): array {
     $fields = parent::baseFieldDefinitions($entity_type);
-    $fields += self::suggestedTopicsFields($entity_type);
 
-    // @todo Can we move this to SuggestedTopicsTrait?
     $fields['field_recommended_topics'] = BaseFieldDefinition::create('suggested_topics_reference')
       ->setName('field_recommended_topics')
       ->setLabel(new TranslatableMarkup('Automatically selected recommendation topics', [], ['context' => 'Recommendations']))
       ->setTargetEntityTypeId('tpr_service')
+      ->setTargetBundle('tpr_service')
       ->setReadonly(TRUE)
       ->setTranslatable(FALSE)
       ->setDisplayOptions('form', [
